+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next
 
+- Replace typecasters with proper type definitions.
+  - Instead of `typecaster(type) { |value, _| ... }` you'll have to use `typecaster(type) { |value| ... }`.
+  - Consequently you can access type definition in typecaster, e.g. `typecaster('Object') { |value| value if value.class < type }`, here `type` comes from type definition.
+
 ## v0.1.1
 
 - Fixed represented error message copying when represented model uses symbols for `message`.